requirements.txt
openai==0.28
langchain
python-dotenv
prettytable